requirements.txt
panel
param
openai
langchain
wikipedia
python-dotenv
requests
langchain-community